+++ adept/container/local.py
@@ -111,8 +111,8 @@ class Local(Container):
 
             self.agent.observe(
                 obs,
-                rewards.to(self.device),
-                terminals.to(self.device),
+                rewards.to(self.device).float(),
+                terminals.to(self.device).float(),
                 infos
             )
